requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
python-dotenv==1.0.1
Werkzeug==3.1.3
